Move `findLanguage` function into helpers.ts
