+++ extensions/ql-vscode/src/run-remote-query.ts
@@ -1,7 +1,7 @@
 import { QuickPickItem, Uri, window } from 'vscode';
 import * as yaml from 'js-yaml';
 import * as fs from 'fs-extra';
-import { getOnDiskWorkspaceFolders, showAndLogErrorMessage, showAndLogInformationMessage } from './helpers';
+import { findLanguage, showAndLogErrorMessage, showAndLogInformationMessage } from './helpers';
 import { Credentials } from './authentication';
 import * as cli from './cli';
 import { logger } from './logging';
@@ -15,37 +15,6 @@ interface Config {
 // Test "controller" repository and workflow.
 const OWNER = 'dsp-testing';
 const REPO = 'qc-controller';
-
-/**
- * Finds the language that a query targets.
- * If it can't be autodetected, prompt the user to specify the language manually.
- */
-export async function findLanguage(
-  cliServer: cli.CodeQLCliServer,
-  queryUri: Uri | undefined
-): Promise<string | undefined> {
-  const uri = queryUri || window.activeTextEditor?.document.uri;
-  if (uri !== undefined) {
-    try {
-      const queryInfo = await cliServer.resolveQueryByLanguage(getOnDiskWorkspaceFolders(), uri);
-      const language = (Object.keys(queryInfo.byLanguage))[0];
-      void logger.log(`Detected query language: ${language}`);
-      return language;
-    } catch (e) {
-      void logger.log('Could not autodetect query language. Select language manually.');
-    }
-  }
-  const availableLanguages = Object.keys(await cliServer.resolveLanguages());
-  const language = await window.showQuickPick(
-    availableLanguages,
-    { placeHolder: 'Select target language for your query', ignoreFocusOut: true }
-  );
-  if (!language) {
-    // This only happens if the user cancels the quick pick.
-    void showAndLogErrorMessage('Language not found. Language must be specified manually.');
-  }
-  return language;
-}
 
 interface RepoListQuickPickItem extends QuickPickItem {
   repoList: string[];
